backend/requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.1
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
click==8.1.6
django-localflavor==4.0
Django==4.2.3
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
dnspython==2.4.1
drf-yasg==1.21.7
eventlet==0.33.3
greenlet==2.0.2
kombu==5.3.1
prompt-toolkit==3.0.39
PyJWT==2.8.0
python-dotenv==1.0.0
python-stdnum==1.18
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
validate-docbr==1.10.0
vine==5.0.0
wcwidth==0.2.6
django-cors-headers==4.2.0
psycopg==3.1.9
Pillow==10.0.0